Move stories to subfolders
- Looks cleaner when having multiple stories and docs
