+++ stories/SpotPrice/AdviceGraph.mdx
@@ -1,0 +1,14 @@
+import { Canvas, Meta, Title } from '@storybook/blocks'
+import * as AdviceGraphStories from './AdviceGraph.stories'
+
+<Meta of={AdviceGraphStories} name="Documentation" />
+
+<Title>Advice Graph Docs</Title>
+
+This component can be used to render the Advice Graph with data from the API.
+
+The default view is as follows:
+
+<Canvas of={AdviceGraphStories.Default} />
+
+## Use Case
